+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `config.apply_configuration()` function is also deprecated.
 - The `theme_style` option is removed. Use `:colorscheme` command instead.
 - The module `plugins.lualine` moved to `util.lualine` and `plugins` module removed.
+- Config Updates
+  - `msg_area_style` -> Removed.
+  - `comment_style` -> `options.styles.comments`
+  - `function_style` -> `options.styles.functions`
+  - `keyword_style` -> `options.styles.keywords`
+  - `variable_style` -> `options.styles.variables`
 
 ### What's New?